feat(shimmer): create a new shimmer for expanded text
