+++ lib/components/shimmers/expanded_image_view.dart
@@ -1,8 +1,9 @@
+import 'package:dalal_street_client/components/shimmers/expanded_text.dart';
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
 import 'package:dalal_street_client/theme/colors.dart';
 import 'package:dalal_street_client/components/shimmers/list_image_view.dart';
-import 'package:dalal_street_client/components/shimmers/text_shimmer.dart';
+import 'package:dalal_street_client/components/shimmers/text_column.dart';
 
 /// A widget to display all shimmers, (useful for testing)
 class AllTheShimmers extends StatelessWidget {
@@ -16,7 +17,7 @@ class AllTheShimmers extends StatelessWidget {
           children: const [
             ListImageViewShimmer(),
             ExpandedImageViewShimmer(),
-            TextShimmer()
+            TextColumn()
           ],
         ),
       ),
@@ -53,67 +54,7 @@ class ExpandedImageViewShimmer extends StatelessWidget {
             const SizedBox(
               height: 15.0,
             ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(0, 0, 5, 0),
-              child: Container(
-                height: 15.0,
-                width: double.infinity,
-                decoration: const BoxDecoration(
-                  color: blurredGray,
-                  borderRadius: BorderRadius.all(
-                    Radius.circular(5.0),
-                  ),
-                ),
-              ),
-            ),
-            const SizedBox(
-              height: 10.0,
-            ),
-            Padding(
-              padding: const EdgeInsets.fromLTRB(0, 0, 5, 0),
-              child: Container(
-                height: 15.0,
-                width: double.infinity,
-                decoration: const BoxDecoration(
-                  color: blurredGray,
-                  borderRadius: BorderRadius.all(
-                    Radius.circular(5.0),
-                  ),
-                ),
-              ),
-            ),
-            const SizedBox(
-              height: 10.0,
-            ),
-            FractionallySizedBox(
-              widthFactor: 0.9,
-              child: Container(
-                height: 15.0,
-                width: double.infinity,
-                decoration: const BoxDecoration(
-                  color: blurredGray,
-                  borderRadius: BorderRadius.all(
-                    Radius.circular(5.0),
-                  ),
-                ),
-              ),
-            ),
-            const SizedBox(
-              height: 10.0,
-            ),
-            FractionallySizedBox(
-              widthFactor: 0.8,
-              child: Container(
-                height: 15.0,
-                width: double.infinity,
-                decoration: const BoxDecoration(
-                  color: blurredGray,
-                  borderRadius: BorderRadius.all(
-                    Radius.circular(5.0),
-                  ),
-                ),
-              ),
-            ),
+            ExpandedText(),
           ],
         ),
       ),
